+updated middleware to check for authentification (WIP) + removed locale and restructured /app router
