+++ app/[workspaceid]/chat/[chatid]/page.tsx
