+++ app/[workspaceid]/layout.tsx
