+++ app/error/page.tsx
@@ -1,0 +1,10 @@
+const AccessDenied = () => {
+  return (
+    <div className="flex h-screen items-center justify-center text-center text-red-500">
+      <h1 className="mb-12">Access Denied</h1>
+      <p>You do not have permission to access this page.</p>
+    </div>
+  )
+}
+
+export default AccessDenied
