+++ app/test/page.tsx
@@ -1,0 +1,9 @@
+const TestMiddlewarePage = () => {
+  return (
+    <div>
+      <h1>Test Middleware Page</h1>
+    </div>
+  )
+}
+
+export default TestMiddlewarePage
