+++ components/chat/chat-ui.tsx
@@ -1,4 +1,4 @@
-import Loading from "@/app/[locale]/loading"
+import Loading from "@/app/loading"
 import { useChatHandler } from "@/components/chat/chat-hooks/use-chat-handler"
 import { ChatbotUIContext } from "@/context/context"
 import { getAssistantToolsByAssistantId } from "@/db/assistant-tools"
